refactor: isolate chatgpt stream projection
